Split procedures bridge code into two files
